Trying to establish a connection between Java and mySQL
Try to follow some tutorial on how to establish a connection between Java and mySQL but it was unsuccessful. The program is running but they cannot found my database local host. Might need to establish some dependencies/ .JAR file or something. Possibly best to create a full java project and follow the tutorial step by step